requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==4.0.0
flask-jwt-extended==4.5.3
pymongo[srv]==4.6.1
python-dotenv==1.0.0
bcrypt==4.0.1
dnspython==2.7.0
PyPDF2==3.0.1
Pillow==11.1.0
pytesseract==0.3.13
numpy==2.2.4
pandas==2.2.3
pdf2image==1.17.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
urllib3==2.2.1
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
idna==3.6
soupsieve==2.5
pydantic
stripe
cryptography==41.0.7
bleach==6.1.0
